Action config view transfer to designer flow - finished
